requirements.txt
groq
python-dotenv
flask
flask-sqlalchemy
